Рефакторинг навигации и структуры экранов в `home`
Этот коммит вносит следующие изменения:

- **Удаление `MainHomeContentScreen.kt`:** Функциональность этого файла была интегрирована в `HomeScreen.kt`.
- **`HomeScreen.kt`:**
    - Преобразован в `Screen` из библиотеки Voyager для навигации.
    - Реализует логику отображения главного экрана, включая информацию о пользователе, текущем стрике, карточки "Завершенные" и "В процессе", а также календарь.
    - Навигация на экраны `ChallengesListScreen` и `ChallengesByDateScreen` теперь осуществляется через Voyager `navigator`.
    - Удален `HomeSubScreen` sealed class, так как навигация теперь управляется Voyager.
    - Кнопка профиля в TopAppBar теперь переключает на `ProfileTab` с помощью `LocalTabNavigator`.
- **`ChallengesListScreen.kt`:**
    - Преобразован в `Screen` из библиотеки Voyager.
    - Принимает `status` (например, "inProgress" или "completed") в качестве параметра для фильтрации челленджей.
    - Кнопка "Назад" теперь использует `navigator.pop()` для возврата на предыдущий экран.
- **`ChallengesByDateScreen.kt`:**
    - Преобразован в `Screen` из библиотеки Voyager.
    - Принимает `date` (LocalDate) в качестве параметра для отображения челленджей на выбранную дату.
    - Кнопка "Назад" теперь использует `navigator.pop()` для возврата на предыдущий экран.
    - Удалено получение `status` из Intent, так как экран теперь фокусируется на отображении по дате, а не по статусу.
    - Компонент `WeekRow` для отображения дней недели вокруг выбранной даты.
    - Добавлен пример использования `ToggleButton` в `Row` для демонстрации группировки кнопок.

В целом, эти изменения упрощают навигацию и структуру экранов в разделе "Дом", используя библиотеку Voyager для более декларативного подхода к управлению экранами.
